Fixed swagger errors
